NWPS-1798-BE-Event-Item-page-Complete-re-vamp
Add / remove fields in CMS as per specs - Event

Note: Location - There is a note in the specs for devs to recommend how this can be setup. A suggestion is to use the ‘Location’ field value to generate a link that opens up to google map. If this is not possible, then please remove the 'View on map' google link from the frontend.