Wsp 256 newsletter setup (#659)
* Update the Newsletter form

Eliminate first name and last name and put only Name field.

* WSP-256-Newsletter-setup

Update the Javascript

* WSP-256-Newsletter-Setup

Update the 'name' argument of some <input> fields in the form required to get the information correctly to the new URL.
Add 'data-id' argument in the <form> tag needed to the info get processed.

* WSP-256-Newsletter-setup

the checkbox consent field is commented because it was requested to remove it to get the form live. But might be needed in the future

* WSP-256-Newsletter-setup

Update the FTL to sent the correct type of information.
Update the Newsletter Signup document to show the correct consent text and URL

* Update Newsletter signup content block

Update the content block newsletter signup

* WSP-256-Newsletter-setup

Delete the last name and first name reference. Only Name is needed.

* [WSP-256] Changes to newsletter ftl,yaml, java files

* Update WSP-256.yml

* [WSP-256] Changes to newsletter ftl,yaml, java files

* [WSP-256] Updating HEE JS assets to include newsletter validation

* [WSP-256] Delete custom workflow WSP-256.yml

---------

Co-authored-by: sankar-shunmuga <s.shunmugasundaram@nhs.net>
Co-authored-by: sankar-shunmuga <95379560+sankar-shunmuga@users.noreply.github.com>